+++ doc/changelog.rst
@@ -5,7 +5,7 @@ Changelog
 1.0.4
 =====
 - Fixed issue where multi-dimensional variables were dropped when converting to xarray.
-- Replaced all print and warning statements with a logger, ``cdflib.logger``.
+- Replaced all print and warning statements with a logger, ``cdflib.logging.logger``.
 
 1.0.3
 =====